pims/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
affine==2.4.0
    # via rasterio
aiofiles==23.1.0
    # via cytomine-pims (setup.py)
amqp==5.1.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
asciitree==0.3.3
    # via zarr
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via rasterio
billiard==3.6.4.0
    # via celery
cachecontrol==0.12.11
    # via cytomine-python-client
celery==5.2.7
    # via cytomine-pims (setup.py)
certifi==2022.12.7
    # via
    #   rasterio
    #   requests
cffi==1.15.1
    # via pyvips
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   rasterio
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   rasterio
click-repl==0.2.0
    # via celery
cligj==0.7.2
    # via rasterio
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
cytomine-python-client==3.0.0
    # via cytomine-pims (setup.py)
entrypoints==0.4
    # via numcodecs
fastapi==0.110.3
    # via
    #   cytomine-pims (setup.py)
    #   fastapi-utils
fastapi-utils==0.7.0
    # via cytomine-pims (setup.py)
fasteners==0.18
    # via zarr
fonttools==4.38.0
    # via matplotlib
future==0.18.3
    # via cytomine-python-client
h11==0.14.0
    # via uvicorn
hiredis==2.2.2
    # via redis
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
imagecodecs==2025.3.30
    # via cytomine-pims (setup.py)
imageio==2.26.0
    # via scikit-image
importlib-metadata==6.0.0
    # via cytomine-pims (setup.py)
kiwisolver==1.4.4
    # via matplotlib
kombu==5.2.4
    # via celery
lazy-loader==0.1
    # via scikit-image
matplotlib==3.7.1
    # via cytomine-pims (setup.py)
msgpack==1.0.4
    # via cachecontrol
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.0
    # via scikit-image
numcodecs==0.11.0
    # via zarr
numpy==1.24.4
    # via
    #   contourpy
    #   cytomine-pims (setup.py)
    #   imagecodecs
    #   imageio
    #   matplotlib
    #   numcodecs
    #   pywavelets
    #   rasterio
    #   scikit-image
    #   scipy
    #   shapely
    #   snuggs
    #   tifffile
    #   zarr
orjson==3.10.3
    # via cytomine-pims (setup.py)
packaging==23.0
    # via
    #   matplotlib
    #   scikit-image
pathvalidate==2.5.2
    # via cytomine-pims (setup.py)
pillow==9.4.0
    # via
    #   cytomine-pims (setup.py)
    #   imageio
    #   matplotlib
    #   scikit-image
pint==0.20.1
    # via cytomine-pims (setup.py)
pkgconfig==1.5.5
    # via pyvips
prompt-toolkit==3.0.38
    # via click-repl
psutil==5.9.8
    # via fastapi-utils
pycparser==2.21
    # via cffi
pydantic==2.7.3
    # via
    #   cytomine-pims (setup.py)
    #   fastapi
    #   fastapi-utils
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-extra-types==2.8.0
    # via cytomine-pims (setup.py)
pydantic-settings==2.3.1
    # via cytomine-pims (setup.py)
pydicom==2.3.1
    # via cytomine-pims (setup.py)
pyparsing==3.0.9
    # via
    #   matplotlib
    #   snuggs
python-dateutil==2.8.2
    # via
    #   cytomine-pims (setup.py)
    #   matplotlib
python-dotenv==1.0.1
    # via
    #   cytomine-pims (setup.py)
    #   pydantic-settings
    #   uvicorn
python-gdcm==3.0.21
    # via cytomine-pims (setup.py)
python-multipart==0.0.9
    # via cytomine-pims (setup.py)
pytz==2022.7.1
    # via celery
pyvips==2.2.3
    # via cytomine-pims (setup.py)
pywavelets==1.4.1
    # via scikit-image
pyyaml==6.0
    # via uvicorn
rasterio==1.3.6
    # via cytomine-pims (setup.py)
redis[hiredis]==5.0.1
    # via cytomine-pims (setup.py)
requests==2.28.2
    # via
    #   cachecontrol
    #   cytomine-python-client
    #   requests-toolbelt
requests-toolbelt==0.10.1
    # via cytomine-python-client
scikit-image==0.20.0
    # via cytomine-pims (setup.py)
scipy==1.9.1
    # via scikit-image
shapely==2.0.1
    # via
    #   cytomine-pims (setup.py)
    #   cytomine-python-client
six==1.16.0
    # via
    #   click-repl
    #   cytomine-python-client
    #   python-dateutil
sniffio==1.3.0
    # via anyio
snuggs==1.4.7
    # via rasterio
starlette==0.37.2
    # via fastapi
tifffile==2023.2.28
    # via
    #   cytomine-pims (setup.py)
    #   scikit-image
typing-extensions==4.12.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via cytomine-pims (setup.py)
urllib3==1.26.14
    # via
    #   cytomine-python-client
    #   requests
uvicorn[standard]==0.29.0
    # via cytomine-pims (setup.py)
uvloop==0.17.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.18.1
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==10.4
    # via uvicorn
zarr==2.14.2
    # via cytomine-pims (setup.py)
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
